Split internal parts from connection.h
